refactor frontend and modelserving
