+++ .dockerignore
@@ -1,8 +1,8 @@
 *.egg-info
-.idea
-.pytest_cache
-data
-.pytest
-venv
-wheels
+*.idea
+*.pytest_cache
+*data
+*.pytest
+*venv
+*wheels
