+++ docker/modelserving.Dockerfile
@@ -1,16 +1,16 @@
 FROM nvcr.io/nvidia/pytorch:23.11-py3
 
-ENV TORCH_HOME=/usr/src/pyproject_microservices/cache
+ENV TORCH_HOME=/usr/src/service_modelserving/cache
 
-WORKDIR /usr/src/pyproject_microservices
+WORKDIR /usr/src/service_modelserving
 
-COPY . .
+COPY ./service_modelserving .
 
 RUN pip install --upgrade pip \
 	&& pip install -e .[all] \
 	&& apt update -y \
 	# && apt -y upgrade \
-	&& apt install -y\
+	&& apt install -y \
 		fonts-humor-sans \
 	# && conda update -y conda \
 	# && while read requirement; do conda install --yes ${requirement}; done < requirements_pytorch.txt \
