--- pyproject_microservices/streamlit/requirements.txt
@@ -1,1 +1,0 @@
-streamlit
