--- pyproject_microservices/streamlit/subpages/__init__.py
@@ -1,2 +1,0 @@
-#! /usr/bin/env python3
-# -*- coding: utf-8 -*-
