+++ service_modelserving/setup.cfg
@@ -4,7 +4,7 @@ parallel = True
 
 [coverage:paths]
 source =
-    pyproject_microservices/
+    service_modelserving/
 
 [coverage:report]
 omit =
@@ -16,7 +16,7 @@ omit =
 
 [coverage:html]
 directory = htmlcov
-title = pyproject_microservices Test Coverage
+title = service_modelserving Test Coverage
 
 # pytest
 [tool:pytest]
@@ -35,5 +35,5 @@ addopts =
     #-k "not file1*"
     --pycodestyle
 testpaths =
-    pyproject_microservices
+    service_modelserving
